refactor(OC): migrate session heart beat to Typescript
